Create 1st step code
this code is the first code that was maid to be presented to the WRO future engineers category. Unlike the code of the first car, this code respect the rule 11.3 since we used a stepper motor to control the steering system. However this code can only be used to pass the round 1 (and 2 if there's 4 rounds) since it doesn't contain the color(s) sensor code