Update 1st step code
This code is the first code to the car that will be presented in the future engineers category of WRO robotics 2023- Lebanon. Unlike the code of the first car, this code respect the rue 11.3 since it contains a stepper motor that controlls the steering system. However this code is only made to pass round 1 ( and 2 if there's 4 rounds) since it doesn't contain a color sensor